requirements.txt
fastapi==0.79.0
requests==2.26.0
uvicorn==0.18.2
gunicorn==20.1.0
uvloop==0.16.0
httptools==0.4.0
pytest-mock==3.8.2